rename : 패키지 GameLauncher -> launcher package로 이동
