+++ src/main/java/baseball/config/AppConfig.java
@@ -2,7 +2,7 @@ package baseball.config;
 
 import baseball.game.BaseballGames;
 import baseball.game.Game;
-import baseball.game.GameLauncher;
+import baseball.launcher.GameLauncher;
 import baseball.service.MessageGenerateService;
 import baseball.service.ValidateJudgeService;
 import baseball.service.impl.MessageGenerateServiceImpl;
