+++ src/main/java/baseball/launcher/GameLauncher.java
@@ -1,5 +1,6 @@
-package baseball.game;
+package baseball.launcher;
 
+import baseball.game.Game;
 import baseball.message.GameMessages;
 import baseball.service.ValidateJudgeService;
 import baseball.vo.BaseballCode;
